fix(tests): 393 tests never ran on this host, and they were hiding two 500s
FS-284b. `make test` reported 1975 passing. It now reports 2709, and the 734 that
appear are the database-backed ones: tenant isolation, RLS, the real migrated
schema. The suite was strongest exactly where nobody could run it.

Ryuk is testcontainers' reaper — a sidecar that bind-mounts the Docker socket to
clean up after a crashed run. On colima that socket is a Lima-forwarded path and
the mount is refused ("operation not supported"), the reaper fails to start,
testcontainers treats that as fatal, and every DB-backed test errors at SETUP.
Errors, not failures, which is why it read as environment noise for long enough
that 393 of them scrolled past in every run this session.

WHAT THEY CAUGHT THE MOMENT THEY RAN — both introduced by this session's own
response_model work, both mine:

  * HealthBandItem.min/max declared `int`. HEALTH_BANDS' top band is
    (…, 100.01), an exclusive upper bound written as a fraction, and pydantic
    refuses a float with a fractional part for an int field. GET
    /dashboard/health/distribution returned 500.
  * HistorianPolicyOut.ingestion_priority declared `str`. It is a numeric
    priority. Every create and update of a retention policy returned 500.

Both are the same class and it is worth naming: the AST sweep compares KEY NAMES
and both keys were named correctly. The unit guards validate against fixtures I
wrote, so they agreed with my own wrong assumption. mypy sees dicts. The only
thing that disagrees with a wrong type is a real row from a real column — which
is precisely what the 393 tests provide and what nothing else in the suite does.

FS-303 in the sprint plan proposed a type-validation sweep to close this class.
It is still worth building, but note what actually caught it first.

COST OF THE FIX: a hard-killed run can now leave a container behind (docker ps;
docker rm). That is a cleanup chore. Silently skipping the database half of the
suite is a correctness risk, and this codebase has already paid for that shape
once — the audit trail passed for months against a pgcrypto extension conftest
created and no migration installed, where "the tests were not wrong about the
code, they were wrong about the database".

Set explicitly via TESTCONTAINERS_RYUK_DISABLED to override either way; the
fixtures only setdefault.

+++ backend/app/api/dashboard_analytics.py
@@ -101,8 +101,12 @@ class AvailabilityTrendResponse(BaseModel):
 
 class HealthBandItem(BaseModel):
     band: str
-    min: int
-    max: int
+    #: FLOATS, not ints. HEALTH_BANDS' top band is (…, 100.01) — an exclusive
+    #: upper bound expressed as a fraction — and pydantic rejects a float with a
+    #: fractional part for an int field, so `int` turned this endpoint into a 500.
+    #: Caught only once the 393 database-backed tests could run.
+    min: float
+    max: float
     count: int
